requirements.txt
azure-identity
azure-mgmt-apimanagement
loguru
pyYaml
pytest
pytest-cov